restored solenoidshifters to its original form, added autosolenoidshifters to do everything I removed
